broke out prior into its own file
