+++ bliss/datasets/simulated.py
@@ -4,7 +4,8 @@ import pytorch_lightning as pl
 import torch
 from torch.utils.data import DataLoader, Dataset, IterableDataset
 
-from bliss.models.decoder import ImageDecoder, ImagePrior
+from bliss.models.prior import ImagePrior
+from bliss.models.decoder import ImageDecoder
 
 # prevent pytorch_lightning warning for num_workers = 0 in dataloaders with IterableDataset
 warnings.filterwarnings(
