+++ docs/_modules/bliss/datasets/simulated.html
@@ -157,7 +157,8 @@
 <span class="kn">import</span> <span class="nn">torch</span>
 <span class="kn">from</span> <span class="nn">torch.utils.data</span> <span class="kn">import</span> <span class="n">DataLoader</span><span class="p">,</span> <span class="n">Dataset</span><span class="p">,</span> <span class="n">IterableDataset</span>
 
-<span class="kn">from</span> <span class="nn">bliss.models.decoder</span> <span class="kn">import</span> <span class="n">ImageDecoder</span><span class="p">,</span> <span class="n">ImagePrior</span>
+<span class="kn">from</span> <span class="nn">bliss.models.prior</span> <span class="kn">import</span> <span class="n">ImagePrior</span>
+<span class="kn">from</span> <span class="nn">bliss.models.decoder</span> <span class="kn">import</span> <span class="n">ImageDecoder</span>
 
 <span class="c1"># prevent pytorch_lightning warning for num_workers = 0 in dataloaders with IterableDataset</span>
 <span class="n">warnings</span><span class="o">.</span><span class="n">filterwarnings</span><span class="p">(</span>
