+++ docs/_modules/bliss/sleep.html
@@ -169,7 +169,8 @@
 <span class="kn">from</span> <span class="nn">torch.distributions</span> <span class="kn">import</span> <span class="n">Normal</span>
 <span class="kn">from</span> <span class="nn">torch.nn</span> <span class="kn">import</span> <span class="n">CrossEntropyLoss</span>
 
-<span class="kn">from</span> <span class="nn">bliss.models.decoder</span> <span class="kn">import</span> <span class="n">ImageDecoder</span><span class="p">,</span> <span class="n">ImagePrior</span>
+<span class="kn">from</span> <span class="nn">bliss.models.prior</span> <span class="kn">import</span> <span class="n">ImagePrior</span>
+<span class="kn">from</span> <span class="nn">bliss.models.decoder</span> <span class="kn">import</span> <span class="n">ImageDecoder</span>
 <span class="kn">from</span> <span class="nn">bliss.models.encoder</span> <span class="kn">import</span> <span class="n">ImageEncoder</span><span class="p">,</span> <span class="n">get_full_params</span><span class="p">,</span> <span class="n">get_is_on_from_n_sources</span>
 <span class="kn">from</span> <span class="nn">bliss.optimizer</span> <span class="kn">import</span> <span class="n">load_optimizer</span>
 <span class="kn">from</span> <span class="nn">bliss.reporting</span> <span class="kn">import</span> <span class="n">DetectionMetrics</span><span class="p">,</span> <span class="n">plot_image_and_locs</span>
