+++ docs/genindex.html
@@ -393,8 +393,6 @@
   <td style="width: 33%; vertical-align: top;"><ul>
       <li><a href="api/bliss.models.html#bliss.models.encoder.ImageEncoder">ImageEncoder (class in bliss.models.encoder)</a>
 </li>
-      <li><a href="api/bliss.models.html#bliss.models.decoder.ImagePrior">ImagePrior (class in bliss.models.decoder)</a>
-</li>
   </ul></td>
 </tr></table>
