+++ docs/searchindex.js
@@ -1,1 +1,1 @@
-Search.setIndex({docnames:["Installation","api/bliss","api/bliss.datasets","api/bliss.models","api/index","index"],envversion:{"sphinx.domains.c":2,"sphinx.domains.changeset":1,"sphinx.domains.citation":1,"sphinx.domains.cpp":3,"sphinx.domains.index":1,"sphinx.domains.javascript":2,"sphinx.domains.math":2,"sphinx.domains.python":2,"sphinx.domains.rst":2,"sphinx.domains.std":2,"sphinx.ext.intersphinx":1,"sphinx.ext.viewcode":1,sphinx:56},filenames:["Installation.rst","api/bliss.rst","api/bliss.datasets.rst","api/bliss.models.rst","api/index.rst","index.rst"],objects:{"bliss.datasets":{sdss:[2,0,0,"-"],simulated:[2,0,0,"-"]},"bliss.datasets.sdss":{SloanDigitalSkySurvey:[2,1,1,""]},"bliss.datasets.simulated":{BlissDataset:[2,1,1,""],SimulatedDataset:[2,1,1,""]},"bliss.datasets.simulated.SimulatedDataset":{test_dataloader:[2,2,1,""],train_dataloader:[2,2,1,""],val_dataloader:[2,2,1,""]},"bliss.models":{decoder:[3,0,0,"-"],encoder:[3,0,0,"-"],fnp:[3,0,0,"-"],galaxy_net:[3,0,0,"-"]},"bliss.models.decoder":{GalaxyTileDecoder:[3,1,1,""],ImageDecoder:[3,1,1,""],ImagePrior:[3,1,1,""],StarTileDecoder:[3,1,1,""],Tiler:[3,1,1,""]},"bliss.models.decoder.GalaxyTileDecoder":{forward:[3,2,1,""]},"bliss.models.decoder.ImageDecoder":{forward:[3,2,1,""]},"bliss.models.decoder.StarTileDecoder":{forward:[3,2,1,""]},"bliss.models.decoder.Tiler":{forward:[3,2,1,""],render_one_source:[3,2,1,""],render_tile:[3,2,1,""]},"bliss.models.encoder":{ConvBlock:[3,1,1,""],EncoderCNN:[3,1,1,""],ImageEncoder:[3,1,1,""],get_images_in_tiles:[3,3,1,""],get_is_on_from_n_sources:[3,3,1,""]},"bliss.models.encoder.ConvBlock":{forward:[3,2,1,""]},"bliss.models.encoder.EncoderCNN":{forward:[3,2,1,""]},"bliss.models.encoder.ImageEncoder":{forward:[3,2,1,""],get_images_in_tiles:[3,2,1,""]},"bliss.models.fnp":{AveragePooler:[3,1,1,""],DepGraph:[3,1,1,""],FNP:[3,1,1,""],MAB:[3,1,1,""],PMA:[3,1,1,""],RepEncoder:[3,1,1,""],SAB:[3,1,1,""],SetPooler:[3,1,1,""]},"bliss.models.fnp.AveragePooler":{forward:[3,2,1,""]},"bliss.models.fnp.FNP":{calc_log_pqz:[3,2,1,""],cov_vencoder:[3,4,1,""],dep_graph:[3,4,1,""],encode:[3,2,1,""],fb_z:[3,4,1,""],forward:[3,2,1,""],label_vdecoder:[3,4,1,""],pooler:[3,4,1,""],prop_vencoder:[3,4,1,""],rep_encoder:[3,4,1,""],trans_cond_y:[3,4,1,""]},"bliss.models.fnp.MAB":{forward:[3,2,1,""]},"bliss.models.fnp.PMA":{forward:[3,2,1,""]},"bliss.models.fnp.RepEncoder":{forward:[3,2,1,""]},"bliss.models.fnp.SAB":{forward:[3,2,1,""]},"bliss.models.fnp.SetPooler":{forward:[3,2,1,""]},"bliss.models.galaxy_net":{CenteredGalaxyDecoder:[3,1,1,""],CenteredGalaxyEncoder:[3,1,1,""],OneCenteredGalaxyAE:[3,1,1,""],OneCenteredGalaxyDecoder:[3,1,1,""],OneCenteredGalaxyEncoder:[3,1,1,""],ResConv2dBlock:[3,1,1,""],ResidualConvBlock:[3,1,1,""]},"bliss.models.galaxy_net.CenteredGalaxyDecoder":{forward:[3,2,1,""]},"bliss.models.galaxy_net.CenteredGalaxyEncoder":{forward:[3,2,1,""]},"bliss.models.galaxy_net.OneCenteredGalaxyAE":{configure_optimizers:[3,2,1,""],forward:[3,2,1,""],generate_latents:[3,2,1,""],latent_dim:[3,4,1,""],main_autoencoder:[3,4,1,""],main_decoder:[3,4,1,""],main_encoder:[3,4,1,""],min_sd:[3,4,1,""],optimizer_step:[3,2,1,""],residual_autoencoder:[3,4,1,""],residual_decoder:[3,4,1,""],residual_delay_n_steps:[3,4,1,""],residual_encoder:[3,4,1,""],slen:[3,4,1,""],test_step:[3,2,1,""],training_step:[3,2,1,""],validation_epoch_end:[3,2,1,""],validation_step:[3,2,1,""]},"bliss.models.galaxy_net.OneCenteredGalaxyDecoder":{forward:[3,2,1,""]},"bliss.models.galaxy_net.OneCenteredGalaxyEncoder":{forward:[3,2,1,""]},"bliss.models.galaxy_net.ResConv2dBlock":{forward:[3,2,1,""]},"bliss.models.galaxy_net.ResidualConvBlock":{forward:[3,2,1,""]},"bliss.sleep":{SleepPhase:[1,1,1,""]},"bliss.sleep.SleepPhase":{configure_optimizers:[1,2,1,""],forward:[1,2,1,""],test_step:[1,2,1,""],training_step:[1,2,1,""],validation_epoch_end:[1,2,1,""],validation_step:[1,2,1,""]},bliss:{generate:[1,0,0,"-"],sleep:[1,0,0,"-"],train:[1,0,0,"-"],tune:[1,0,0,"-"]}},objnames:{"0":["py","module","Python module"],"1":["py","class","Python class"],"2":["py","method","Python method"],"3":["py","function","Python function"],"4":["py","attribute","Python attribute"]},objtypes:{"0":"py:module","1":"py:class","2":"py:method","3":"py:function","4":"py:attribute"},terms:{"001":3,"00825":3,"08324":3,"10000":3,"1000000":3,"128":3,"160":3,"1810":3,"1906":3,"269":2,"3900":2,"500":3,"686":3,"boolean":3,"case":[2,3],"class":[1,2,3],"default":3,"final":0,"float":3,"function":[3,5],"import":[1,4],"int":3,"return":[2,3],"true":[2,3],"while":3,For:[0,2,3],The:[2,3],Then:0,There:2,These:3,a_in:3,abl:0,abov:2,abs:3,absolut:0,accumul:3,accumulate_grad_batch:3,accur:5,add:[2,3],added:3,adjust:3,advanc:3,afterward:3,algorithm:5,all:[2,3],allow:3,allow_pad:3,alpha:3,alreadi:5,also:2,altern:3,although:3,amlab:3,amort:5,amount:3,amp:3,amsterdam:3,ani:[2,3],annotate_prob:1,anoth:3,api:5,appli:3,applic:3,approxim:5,apt:0,arbitrari:2,architectur:3,arg:2,argument:[2,3],arrai:3,art:5,arxiv:3,assign:2,astronom:[3,5],attent:3,attribut:3,author:3,auto:5,autoencod:[3,5],autoencoder_ckpt:3,averag:3,averagepool:3,avoid:3,background:3,background_valu:3,backward:3,band:2,base:[3,5],batch:[1,2,3],batch_cifar:2,batch_idx:[1,3],batch_mnist:2,batch_siz:2,batchnorm:3,batchsiz:3,been:3,befor:3,below:0,better:3,between:3,bia:3,bin:2,bit:3,blend:5,bliss:0,bliss_hom:0,blissdataset:2,block:3,bool:[1,3],border_pad:3,both:3,bug:3,bulg:5,calc_log_pqz:3,calcul:3,calibr:5,call:[0,3],camcol:2,can:[0,2,3],care:3,catsim:[1,5],center:3,center_subpixel:2,centeredgalaxydecod:3,centeredgalaxyencod:3,cfg:1,chang:3,channel:3,cifar:2,cifar_load:2,clone:0,closur:3,cnn:3,collect:2,com:[0,3],come:3,command:0,compos:2,comput:3,configur:[1,3],configure_optim:[1,3],consist:3,construct:1,constructor:3,contain:[0,1],conv2d:3,convblock:3,convolut:3,correct:2,correctli:0,could:3,cov_vencod:3,creat:[0,2,3],current:3,data:[1,2],dataload:[2,3],dataloader_idx:2,dataset:4,deblend:[3,5],deblendingstarfield:5,decod:1,def:[2,3],defin:3,dep_graph:3,depend:[0,3],depgraph:3,detect:5,dev:0,develop:5,dict:[1,2,3],differ:3,dilat:3,dim:3,dim_in:3,dim_k:3,dim_out:3,dim_q:3,dim_rep_r:3,dim_u:3,dim_v:3,dim_z:3,dimens:3,directori:0,discrimin:3,disk:5,distribut:[2,3],divid:3,don:2,download:2,downsampl:3,dropout:3,dure:3,each:[2,3],either:3,embed:3,encod:[1,5],encodercnn:3,encourag:3,end:3,ensur:3,entir:[3,5],entri:3,environ:0,epoch:3,estim:5,everi:3,everyth:0,exampl:[0,1,2,3],expand_factor:3,expanded_psf:3,f_max:3,f_min:3,fals:[1,2,3],fb_z:3,featur:3,few:3,fftw:0,field:[2,5],first:[0,3],fit:[1,2,5],flatten:3,flexibl:5,flux:3,fnp:1,follow:[2,3],form:3,former:3,forward:[1,3],framework:1,free:3,from:[1,2,3,5],full:3,g_in:3,gal_slen:3,galaxi:3,galaxy_bool:3,galaxy_net:1,galaxy_param:3,galaxytiledecod:3,galsim:[0,3],gan:3,gener:[2,3,4],generate_lat:3,get:[0,3],get_images_in_til:3,get_is_on_from_n_sourc:3,git:0,github:[0,3],given:3,global_step:3,graph:3,graphic:3,group:3,guid:5,happen:2,hardwar:2,have:[2,3],here:[2,3],hidden:3,hook:3,how:3,howev:2,http:[0,3],ignor:3,imag:[1,3,5],image_ptil:[1,3],imagedecod:3,imageencod:3,imageprior:3,implement:[1,2,3],in_channel:3,includ:5,index:3,indic:3,induc:3,infer:[3,5],input:3,instal:5,instanc:3,instead:3,integ:2,interest:5,invari:3,isol:3,its:[3,5],jointli:5,juho:3,kei:3,kernel_s:3,kwarg:[1,2],label:3,label_vdecod:3,larger:3,latent:3,latent_dim:3,latents_fil:3,latter:3,layer:3,lbfg:3,learn:3,learning_r:3,lee:3,len:3,length:3,libfftw3:0,librari:0,licens:3,lightn:[1,2,3],like:3,likelihood:3,line:1,list:[2,3],loader:2,loader_a:2,loader_b:2,loader_n:2,loc:3,loc_max:3,loc_min:3,local:3,locat:3,log:3,loop:3,loss:3,lr_scale:3,mab:3,made:3,main:3,main_autoencod:3,main_decod:3,main_encod:3,make:3,mani:3,map:5,match:[2,3],matric:3,matrix:3,max_detect:3,max_num_star:3,max_sourc:3,maximum:3,mean_sourc:3,measur:5,method:[1,2,3,5],metric:4,middl:3,min:3,min_sd:3,min_sourc:3,minim:3,minimum:3,mit:3,mnist:2,mnist_load:2,mode:3,model:[4,5],modul:4,more:[2,3],morpholog:5,mostli:3,multihead:3,multipl:[2,3],n_band:3,n_batch:3,n_galaxy_param:3,n_latent_batch:3,n_layer:3,n_ptile:3,n_sourc:3,nativ:3,necessari:2,need:[0,2,3],neg:3,network:5,neural:[3,5],non:3,none:[1,3],normal:2,note:3,now:5,num_head:3,num_se:3,number:3,object:3,often:3,on_tpu:3,onc:3,one:[2,3],onecenteredgalaxya:3,onecenteredgalaxydecod:3,onecenteredgalaxyencod:3,onli:[0,2,3],opt:3,optim:[1,3],optimizer_closur:3,optimizer_idx:[1,3],optimizer_param:[1,3],optimizer_step:3,option:[1,3],order:2,org:3,origin:3,out_channel:3,outout:3,output:[1,3],output_pad:3,overrid:3,overridden:3,overwrite_cach:2,overwrite_fits_cach:2,own:3,packag:4,pad:3,padding_mod:3,page:2,param:3,param_group:3,paramet:[1,3,5],paramref:2,part:3,pass:3,path:[0,2],pattern:2,per:3,perform:3,permut:3,phase:[1,3],pleas:2,plot:4,pma:3,poetri:0,point:3,pool:3,pooler:3,pooling_lay:3,posit:2,posterior:5,postiv:2,predecessor:5,predict:3,prepar:2,prepare_data:2,present:3,prior:[1,3],prob:0,prob_galaxi:3,probabilist:3,probabl:3,problem:3,procedur:[3,5],process:[2,3],prop_vencod:3,proport:3,provid:[3,5],psf:[3,5],psf_params_fil:3,psf_slen:3,pt_file:2,ptile:3,ptile_slen:3,python:1,pytorch:[1,2,3],pytorch_lightn:[1,2],queri:3,rate:3,recip:3,recommend:2,reconstruct:3,refer:3,regist:3,regular:3,reload:2,reload_dataloaders_every_n_epoch:2,relu:3,render:3,render_one_sourc:3,render_til:3,rep_encod:3,rep_r:3,repeat:3,repencod:3,repo:5,repres:3,represent:3,requir:3,resconv2dblock:3,residu:3,residual_autoencod:3,residual_decod:3,residual_delay_n_step:3,residual_encod:3,residualconvblock:3,root:2,run:[0,2,3],sab:3,same:3,sampl:[2,3],sample_a:3,sample_g:3,sampler:[2,3],save:2,scalabl:5,schedul:3,script:1,sdss:1,sdss_band:3,sdss_dir:2,second:3,see:[2,3],self:[2,3],sequenc:2,set:[0,1,2,3],set_transform:3,setpool:3,setup:2,shape:[3,5],share:3,shell:0,should:[0,1,3],show:3,shown:3,shuffl:2,side:3,signific:3,silent:3,similar:3,simpli:0,simul:1,simulateddataset:2,sinc:3,singl:[2,3],size:3,slack:1,sleep:[4,5],sleepphas:1,slen:3,slide:3,sloandigitalskysurvei:2,smaller:3,sourc:[1,2,3],spatial_dropout:3,specif:3,specifi:2,split:2,squeeze_out:3,st_numhead:3,stack:3,stampsiz:[2,3],star:3,star_bool:3,starnet:5,start:3,startiledecod:3,state:[2,5],step:[1,3],str:3,strategi:3,stride:3,structur:3,subclass:3,sudo:0,sum:3,summari:1,suppos:3,sure:3,survei:5,take:3,telescop:5,temperatur:3,tensor:[2,3],test:[2,3,5],test_dataload:2,test_step:[1,2,3],than:3,thei:3,them:[2,3],thi:[0,1,2,3],thing:3,through:[3,5],tile:[1,3],tile_n_sourc:[1,3],tile_slen:3,tile_weight:3,tile_width:3,tiler:3,tiles_per_batch:3,time:3,torch:[2,3],totensor:2,tpu:3,train:[2,3,4,5],train_dataload:2,trainer:[1,2,3],training_step:[1,3],trans_cond_i:3,transform:[2,3],tune:4,tupl:3,two:3,type:3,uncertainti:5,union:3,unix:0,unless:2,updat:3,use:[0,1,2,3],use_u_diff:3,use_x:3,used:[0,3],user:1,uses:5,using:[1,3],using_lbfg:3,using_native_amp:3,util:2,vae:5,val_dataload:2,valid:[2,3],validation_epoch_end:[1,3],validation_step:[1,2,3],valu:3,variabl:[0,3],variat:5,wai:3,wake:[4,5],want:3,warm:3,when:3,where:2,whether:3,which:[0,1,2,3,5],whose:3,window:3,within:3,without:3,won:3,work:[3,5],y_encod:3,you:[0,2,3],yourself:2,yr_encod:3,zero:3,zero_grad:3},titles:["Installation Guide","bliss package","bliss.datasets package","bliss.models package","Bliss API","Bayesian Light Source Separator (BLISS)"],titleterms:{api:4,bayesian:5,bliss:[1,2,3,4,5],catsim:2,content:5,dataset:[1,2],decod:3,encod:3,fnp:3,galaxi:5,galaxy_net:3,gener:1,guid:0,instal:0,latest:5,light:5,metric:1,model:[1,3],modul:[1,2,3],packag:[1,2,3],plot:1,sdss:2,separ:5,simul:2,sleep:1,sourc:5,star:5,tabl:5,train:1,tune:1,updat:5,wake:1}})
+Search.setIndex({docnames:["Installation","api/bliss","api/bliss.datasets","api/bliss.models","api/index","index"],envversion:{"sphinx.domains.c":2,"sphinx.domains.changeset":1,"sphinx.domains.citation":1,"sphinx.domains.cpp":3,"sphinx.domains.index":1,"sphinx.domains.javascript":2,"sphinx.domains.math":2,"sphinx.domains.python":2,"sphinx.domains.rst":2,"sphinx.domains.std":2,"sphinx.ext.intersphinx":1,"sphinx.ext.viewcode":1,sphinx:56},filenames:["Installation.rst","api/bliss.rst","api/bliss.datasets.rst","api/bliss.models.rst","api/index.rst","index.rst"],objects:{"bliss.datasets":{sdss:[2,0,0,"-"],simulated:[2,0,0,"-"]},"bliss.datasets.sdss":{SloanDigitalSkySurvey:[2,1,1,""]},"bliss.datasets.simulated":{BlissDataset:[2,1,1,""],SimulatedDataset:[2,1,1,""]},"bliss.datasets.simulated.SimulatedDataset":{test_dataloader:[2,2,1,""],train_dataloader:[2,2,1,""],val_dataloader:[2,2,1,""]},"bliss.models":{decoder:[3,0,0,"-"],encoder:[3,0,0,"-"],fnp:[3,0,0,"-"],galaxy_net:[3,0,0,"-"]},"bliss.models.decoder":{GalaxyTileDecoder:[3,1,1,""],ImageDecoder:[3,1,1,""],StarTileDecoder:[3,1,1,""],Tiler:[3,1,1,""]},"bliss.models.decoder.GalaxyTileDecoder":{forward:[3,2,1,""]},"bliss.models.decoder.ImageDecoder":{forward:[3,2,1,""]},"bliss.models.decoder.StarTileDecoder":{forward:[3,2,1,""]},"bliss.models.decoder.Tiler":{forward:[3,2,1,""],render_one_source:[3,2,1,""],render_tile:[3,2,1,""]},"bliss.models.encoder":{ConvBlock:[3,1,1,""],EncoderCNN:[3,1,1,""],ImageEncoder:[3,1,1,""],get_images_in_tiles:[3,3,1,""],get_is_on_from_n_sources:[3,3,1,""]},"bliss.models.encoder.ConvBlock":{forward:[3,2,1,""]},"bliss.models.encoder.EncoderCNN":{forward:[3,2,1,""]},"bliss.models.encoder.ImageEncoder":{forward:[3,2,1,""],get_images_in_tiles:[3,2,1,""]},"bliss.models.fnp":{AveragePooler:[3,1,1,""],DepGraph:[3,1,1,""],FNP:[3,1,1,""],MAB:[3,1,1,""],PMA:[3,1,1,""],RepEncoder:[3,1,1,""],SAB:[3,1,1,""],SetPooler:[3,1,1,""]},"bliss.models.fnp.AveragePooler":{forward:[3,2,1,""]},"bliss.models.fnp.FNP":{calc_log_pqz:[3,2,1,""],cov_vencoder:[3,4,1,""],dep_graph:[3,4,1,""],encode:[3,2,1,""],fb_z:[3,4,1,""],forward:[3,2,1,""],label_vdecoder:[3,4,1,""],pooler:[3,4,1,""],prop_vencoder:[3,4,1,""],rep_encoder:[3,4,1,""],trans_cond_y:[3,4,1,""]},"bliss.models.fnp.MAB":{forward:[3,2,1,""]},"bliss.models.fnp.PMA":{forward:[3,2,1,""]},"bliss.models.fnp.RepEncoder":{forward:[3,2,1,""]},"bliss.models.fnp.SAB":{forward:[3,2,1,""]},"bliss.models.fnp.SetPooler":{forward:[3,2,1,""]},"bliss.models.galaxy_net":{CenteredGalaxyDecoder:[3,1,1,""],CenteredGalaxyEncoder:[3,1,1,""],OneCenteredGalaxyAE:[3,1,1,""],OneCenteredGalaxyDecoder:[3,1,1,""],OneCenteredGalaxyEncoder:[3,1,1,""],ResConv2dBlock:[3,1,1,""],ResidualConvBlock:[3,1,1,""]},"bliss.models.galaxy_net.CenteredGalaxyDecoder":{forward:[3,2,1,""]},"bliss.models.galaxy_net.CenteredGalaxyEncoder":{forward:[3,2,1,""]},"bliss.models.galaxy_net.OneCenteredGalaxyAE":{configure_optimizers:[3,2,1,""],forward:[3,2,1,""],generate_latents:[3,2,1,""],latent_dim:[3,4,1,""],main_autoencoder:[3,4,1,""],main_decoder:[3,4,1,""],main_encoder:[3,4,1,""],min_sd:[3,4,1,""],optimizer_step:[3,2,1,""],residual_autoencoder:[3,4,1,""],residual_decoder:[3,4,1,""],residual_delay_n_steps:[3,4,1,""],residual_encoder:[3,4,1,""],slen:[3,4,1,""],test_step:[3,2,1,""],training_step:[3,2,1,""],validation_epoch_end:[3,2,1,""],validation_step:[3,2,1,""]},"bliss.models.galaxy_net.OneCenteredGalaxyDecoder":{forward:[3,2,1,""]},"bliss.models.galaxy_net.OneCenteredGalaxyEncoder":{forward:[3,2,1,""]},"bliss.models.galaxy_net.ResConv2dBlock":{forward:[3,2,1,""]},"bliss.models.galaxy_net.ResidualConvBlock":{forward:[3,2,1,""]},"bliss.sleep":{SleepPhase:[1,1,1,""]},"bliss.sleep.SleepPhase":{configure_optimizers:[1,2,1,""],forward:[1,2,1,""],test_step:[1,2,1,""],training_step:[1,2,1,""],validation_epoch_end:[1,2,1,""],validation_step:[1,2,1,""]},bliss:{generate:[1,0,0,"-"],sleep:[1,0,0,"-"],train:[1,0,0,"-"],tune:[1,0,0,"-"]}},objnames:{"0":["py","module","Python module"],"1":["py","class","Python class"],"2":["py","method","Python method"],"3":["py","function","Python function"],"4":["py","attribute","Python attribute"]},objtypes:{"0":"py:module","1":"py:class","2":"py:method","3":"py:function","4":"py:attribute"},terms:{"001":3,"00825":3,"08324":3,"128":3,"1810":3,"1906":3,"269":2,"3900":2,"500":3,"686":3,"boolean":3,"case":[2,3],"class":[1,2,3],"default":3,"final":0,"float":3,"function":[3,5],"import":[1,4],"int":3,"return":[2,3],"true":[2,3],"while":3,For:[0,2,3],The:[2,3],Then:0,There:2,These:3,a_in:3,abl:0,abov:2,abs:3,absolut:0,accumul:3,accumulate_grad_batch:3,accur:5,add:[2,3],added:3,adjust:3,advanc:3,afterward:3,algorithm:5,all:[2,3],allow:3,allow_pad:3,alreadi:5,also:2,altern:3,although:3,amlab:3,amort:5,amount:3,amp:3,amsterdam:3,ani:[2,3],annotate_prob:1,anoth:3,api:5,appli:3,applic:3,approxim:5,apt:0,arbitrari:2,architectur:3,arg:2,argument:[2,3],arrai:3,art:5,arxiv:3,assign:2,astronom:[3,5],attent:3,attribut:3,author:3,auto:5,autoencod:[3,5],autoencoder_ckpt:3,averag:3,averagepool:3,avoid:3,background:3,background_valu:3,backward:3,band:2,base:[3,5],batch:[1,2,3],batch_cifar:2,batch_idx:[1,3],batch_mnist:2,batch_siz:2,batchnorm:3,batchsiz:3,been:3,befor:3,below:0,better:3,between:3,bia:3,bin:2,bit:3,blend:5,bliss:0,bliss_hom:0,blissdataset:2,block:3,bool:[1,3],border_pad:3,both:3,bug:3,bulg:5,calc_log_pqz:3,calcul:3,calibr:5,call:[0,3],camcol:2,can:[0,2,3],care:3,catsim:[1,5],center:3,center_subpixel:2,centeredgalaxydecod:3,centeredgalaxyencod:3,cfg:1,chang:3,channel:3,cifar:2,cifar_load:2,clone:0,closur:3,cnn:3,collect:2,com:[0,3],come:3,command:0,compos:2,comput:3,configur:[1,3],configure_optim:[1,3],consist:3,construct:1,constructor:3,contain:[0,1],conv2d:3,convblock:3,convolut:3,correct:2,correctli:0,could:3,cov_vencod:3,creat:[0,2,3],current:3,data:[1,2],dataload:[2,3],dataloader_idx:2,dataset:4,deblend:[3,5],deblendingstarfield:5,decod:1,def:[2,3],defin:3,dep_graph:3,depend:[0,3],depgraph:3,detect:5,dev:0,develop:5,dict:[1,2,3],differ:3,dilat:3,dim:3,dim_in:3,dim_k:3,dim_out:3,dim_q:3,dim_rep_r:3,dim_u:3,dim_v:3,dim_z:3,dimens:3,directori:0,discrimin:3,disk:5,distribut:[2,3],divid:3,don:2,download:2,downsampl:3,dropout:3,dure:3,each:[2,3],either:3,embed:3,encod:[1,5],encodercnn:3,encourag:3,end:3,ensur:3,entir:[3,5],entri:3,environ:0,epoch:3,estim:5,everi:3,everyth:0,exampl:[0,1,2,3],expand_factor:3,expanded_psf:3,fals:[1,2,3],fb_z:3,featur:3,few:3,fftw:0,field:[2,5],first:[0,3],fit:[1,2,5],flatten:3,flexibl:5,flux:3,fnp:1,follow:[2,3],form:3,former:3,forward:[1,3],framework:1,free:3,from:[1,2,3,5],full:3,g_in:3,gal_slen:3,galaxi:3,galaxy_bool:3,galaxy_net:1,galaxy_param:3,galaxytiledecod:3,galsim:[0,3],gan:3,gener:[2,3,4],generate_lat:3,get:[0,3],get_images_in_til:3,get_is_on_from_n_sourc:3,git:0,github:[0,3],given:3,global_step:3,graph:3,graphic:3,group:3,guid:5,happen:2,hardwar:2,have:[2,3],here:[2,3],hidden:3,hook:3,how:3,howev:2,http:[0,3],ignor:3,imag:[1,3,5],image_ptil:[1,3],imagedecod:3,imageencod:3,implement:[1,2,3],in_channel:3,includ:5,index:3,indic:3,induc:3,infer:[3,5],input:3,instal:5,instanc:3,instead:3,integ:2,interest:5,invari:3,isol:3,its:[3,5],jointli:5,juho:3,kei:3,kernel_s:3,kwarg:[1,2],label:3,label_vdecod:3,larger:3,latent:3,latent_dim:3,latter:3,layer:3,lbfg:3,learn:3,learning_r:3,lee:3,len:3,length:3,libfftw3:0,librari:0,licens:3,lightn:[1,2,3],like:3,likelihood:3,line:1,list:[2,3],loader:2,loader_a:2,loader_b:2,loader_n:2,loc:3,local:3,locat:3,log:3,loop:3,loss:3,lr_scale:3,mab:3,made:3,main:3,main_autoencod:3,main_decod:3,main_encod:3,make:3,mani:3,map:5,match:[2,3],matric:3,matrix:3,max_detect:3,max_num_star:3,max_sourc:3,maximum:3,measur:5,method:[1,2,3,5],metric:4,middl:3,min:3,min_sd:3,minim:3,minimum:3,mit:3,mnist:2,mnist_load:2,mode:3,model:[4,5],modul:4,more:[2,3],morpholog:5,mostli:3,multihead:3,multipl:[2,3],n_band:3,n_batch:3,n_galaxy_param:3,n_layer:3,n_ptile:3,n_sourc:3,nativ:3,necessari:2,need:[0,2,3],neg:3,network:5,neural:[3,5],non:3,none:[1,3],normal:2,note:3,now:5,num_head:3,num_se:3,number:3,object:3,often:3,on_tpu:3,onc:3,one:[2,3],onecenteredgalaxya:3,onecenteredgalaxydecod:3,onecenteredgalaxyencod:3,onli:[0,2,3],opt:3,optim:[1,3],optimizer_closur:3,optimizer_idx:[1,3],optimizer_param:[1,3],optimizer_step:3,option:[1,3],order:2,org:3,origin:3,out_channel:3,outout:3,output:[1,3],output_pad:3,overrid:3,overridden:3,overwrite_cach:2,overwrite_fits_cach:2,own:3,packag:4,pad:3,padding_mod:3,page:2,param:3,param_group:3,paramet:[1,3,5],paramref:2,part:3,pass:3,path:[0,2],pattern:2,per:3,perform:3,permut:3,phase:[1,3],pleas:2,plot:4,pma:3,poetri:0,point:3,pool:3,pooler:3,pooling_lay:3,posit:2,posterior:5,postiv:2,predecessor:5,predict:3,prepar:2,prepare_data:2,present:3,prior:[1,3],prob:0,prob_galaxi:3,probabilist:3,probabl:3,problem:3,procedur:[3,5],process:[2,3],prop_vencod:3,proport:3,provid:[3,5],psf:[3,5],psf_params_fil:3,psf_slen:3,pt_file:2,ptile:3,ptile_slen:3,python:1,pytorch:[1,2,3],pytorch_lightn:[1,2],queri:3,rate:3,recip:3,recommend:2,reconstruct:3,refer:3,regist:3,regular:3,reload:2,reload_dataloaders_every_n_epoch:2,relu:3,render:3,render_one_sourc:3,render_til:3,rep_encod:3,rep_r:3,repeat:3,repencod:3,repo:5,repres:3,represent:3,requir:3,resconv2dblock:3,residu:3,residual_autoencod:3,residual_decod:3,residual_delay_n_step:3,residual_encod:3,residualconvblock:3,root:2,run:[0,2,3],sab:3,same:3,sampl:[2,3],sample_a:3,sample_g:3,sampler:[2,3],save:2,scalabl:5,schedul:3,script:1,sdss:1,sdss_band:3,sdss_dir:2,second:3,see:[2,3],self:[2,3],sequenc:2,set:[0,1,2,3],set_transform:3,setpool:3,setup:2,shape:[3,5],share:3,shell:0,should:[0,1,3],show:3,shown:3,shuffl:2,side:3,signific:3,silent:3,similar:3,simpli:0,simul:1,simulateddataset:2,sinc:3,singl:[2,3],size:3,slack:1,sleep:[4,5],sleepphas:1,slen:3,slide:3,sloandigitalskysurvei:2,smaller:3,sourc:[1,2,3],spatial_dropout:3,specif:3,specifi:2,split:2,squeeze_out:3,st_numhead:3,stack:3,stampsiz:[2,3],star:3,star_bool:3,starnet:5,start:3,startiledecod:3,state:[2,5],step:[1,3],str:3,strategi:3,stride:3,structur:3,subclass:3,sudo:0,sum:3,summari:1,suppos:3,sure:3,survei:5,take:3,telescop:5,temperatur:3,tensor:[2,3],test:[2,3,5],test_dataload:2,test_step:[1,2,3],than:3,thei:3,them:[2,3],thi:[0,1,2,3],thing:3,through:[3,5],tile:[1,3],tile_n_sourc:[1,3],tile_slen:3,tile_weight:3,tile_width:3,tiler:3,tiles_per_batch:3,time:3,torch:[2,3],totensor:2,tpu:3,train:[2,3,4,5],train_dataload:2,trainer:[1,2,3],training_step:[1,3],trans_cond_i:3,transform:[2,3],tune:4,tupl:3,two:3,type:3,uncertainti:5,union:3,unix:0,unless:2,updat:3,use:[0,1,2,3],use_u_diff:3,use_x:3,used:[0,3],user:1,uses:5,using:[1,3],using_lbfg:3,using_native_amp:3,util:2,vae:5,val_dataload:2,valid:[2,3],validation_epoch_end:[1,3],validation_step:[1,2,3],valu:3,variabl:[0,3],variat:5,wai:3,wake:[4,5],want:3,warm:3,when:3,where:2,whether:3,which:[0,1,2,3,5],whose:3,window:3,within:3,without:3,won:3,work:[3,5],y_encod:3,you:[0,2,3],yourself:2,yr_encod:3,zero:3,zero_grad:3},titles:["Installation Guide","bliss package","bliss.datasets package","bliss.models package","Bliss API","Bayesian Light Source Separator (BLISS)"],titleterms:{api:4,bayesian:5,bliss:[1,2,3,4,5],catsim:2,content:5,dataset:[1,2],decod:3,encod:3,fnp:3,galaxi:5,galaxy_net:3,gener:1,guid:0,instal:0,latest:5,light:5,metric:1,model:[1,3],modul:[1,2,3],packag:[1,2,3],plot:1,sdss:2,separ:5,simul:2,sleep:1,sourc:5,star:5,tabl:5,train:1,tune:1,updat:5,wake:1}})
